pyproject.toml
[project]
name = "llm-kg-extraction"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai (>=1.76.2,<2.0.0)",
    "pymupdf (>=1.25.5,<2.0.0)",
    "networkx (>=3.4.2,<4.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "pyvis (>=0.3.2,<0.4.0)",
]

[tool.poetry]
packages = [{include = "llm_kg_extraction", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
